requirements.txt
botocore
jsonschema
configparser
click
pyyaml
